// shared/src/constants/species.d.ts
export interface MudCrabSpecies {
    id: string;
    scientificName: string;
    commonName: string;
    shortName: string;
}
export declare const MUD_CRAB_SPECIES: MudCrabSpecies[];
export declare const DEFAULT_PAGE_LIMIT = 20;
export declare const MAX_PAGE_LIMIT = 100;
//# sourceMappingURL=species.d.ts.map