// shared/src/constants/regions.d.ts
export type MalaysianState = 'Johor' | 'Kedah' | 'Kelantan' | 'Melaka' | 'Negeri Sembilan' | 'Pahang' | 'Pulau Pinang' | 'Perak' | 'Perlis' | 'Selangor' | 'Terengganu' | 'Sabah' | 'Sarawak' | 'Kuala Lumpur' | 'Labuan' | 'Putrajaya';
export declare const MALAYSIAN_STATES: MalaysianState[];
export declare const COASTAL_STATES: MalaysianState[];
export declare const MALAYSIA_BOUNDS: {
    north: number;
    south: number;
    east: number;
    west: number;
    center: {
        lat: number;
        lng: number;
    };
};
//# sourceMappingURL=regions.d.ts.map